integration_2022_wk48

See merge request oai/openairinterface5g!1845

!1781 Add 4G MBMS tests to RAN-Container-Parent
!1803 Improve dot_product function
!1836 Fix OPENAIR_DIR in CMakeLists.txt
!1840 PHR fix: re-compute TBS after reducing PRBs or MCS
!1841 Tpool: Unlock after broadcast in abortNotifiedFIFO()
!1843 NR DU: nr rrc: hotfix: set default priority correctly
!1763 Remove FlexRAN
!1789 SRB0 via RLC
!1834 cleanup of gNB PHY structures
!1844 RU: stop RF after all threads finished
!1811 NR UE improve TX thread
!1807 Harmonization of RRC SearchSpace configuration
!1819 ldpc offload: initialize R

+++ openair2/RRC/NR/MESSAGES/asn1_msg.h
@@ -72,8 +72,8 @@ uint8_t do_SIB23_NR(rrc_gNB_carrier_data_t *carrier,
 void do_SpCellConfig(gNB_RRC_INST *rrc,
                       struct NR_SpCellConfig  *spconfig);
 
-uint8_t do_RRCReject(uint8_t Mod_id,
-                     uint8_t *const buffer);
+int do_RRCReject(uint8_t Mod_id,
+                 uint8_t *const buffer);
 
 void fill_initial_SpCellConfig(int uid,
                                NR_SpCellConfig_t *SpCellConfig,
